encode(h265): high-bit-depth streams — Main 10, and Main 12 / Rext deeper
The H.265 encoder refused bit_depth > 8 because `H265Encoder` held its
reconstructions and references as `u8` planes; everything below it —
IntraPicture, InterPicture, the deblocker, SAO, the DSP tables — was already
generic over `Sample`. This makes the top generic too, the way the decoder
is at its NAL layer: `H265Encoder` is now a face over `Core<S>`, built as
`Core<u8>` at 8 bits and `Core<u16>` above, and every kernel the deep path
runs is the decoder's own 16-bit one (`HevcDsp::<S>::new(cpu)`, SIMD tiers
included). Pictures cross the face as bytes in `Picture::into_packed`'s
layout — little-endian u16 above 8 bits, `frame_bytes()` doubles — and a
source sample above the declared depth refuses by name.

What follows the depth, and where:
- SPS `bit_depth_luma/chroma_minus8` (one depth, both components, monochrome
  included); the PTL already claimed Main 10 / Rext by depth.
- `sao_offset_abs` cMax: `(1 << (Min(bitDepth, 10) - 5)) - 1`, 31 from 10 up,
  in the writer as in the decision (`write_sao_for` had 8 hard-coded).
  `log2_sao_offset_scale` stays 0 (no PPS range extension).
- Qp'Y = QpY + QpBdOffsetY inside the scaling (already in code_luma_tb /
  code_residual_cu), chroma through the decoder's `chroma_qp` with the
  -QpBdOffsetC clamp, CABAC init at Clip3(0,51,SliceQpY), deblock tc/beta
  scaled inside `deblock_rows` by the depth it is handed.
- The syntax layer's quantiser widened to i32 (`SliceHeader::qp`,
  `write_pps`, `write_slice_header`): the range is -QpBdOffsetY..=51 and a
  header test now reaches its floor through the production writer. The
  encoder's Config keeps `ConstantQp(u8)`, 0..=51 at every depth.
- Lagrangians scaled by 2^(bd-8) against SATD and 2^(2(bd-8)) against SSD —
  HM's DISTORTION_PRECISION_ADJUSTMENT, applied on the rate side so every
  8-bit cost is multiplied by exactly 1.0.

8-bit output is byte-identical: all 109 HEVC gate cells (9 clips x 12
configs + vbv) have the same stream md5, size and recon md5 before and
after (scratch hevc10_md5_{before,after}.txt, diff empty).

Verification added:
- make_encode_sources.sh: 420p10 detail/motion, 422p10, 444p10, 420p12
  clips with two (four) bits of geq noise per sample AFTER the depth
  conversion, so every low-bit pattern is present (probed: 4/4 and 16/16).
- verify_encode.sh: format token `<chroma>p<depth>` split into --format
  and --depth, ffpix -> yuv420p10le/gray10le, depth-aware PSNR, 8-bit rows
  confined to 8-bit clips, 11 new rows (hevc10-* @p10: intra, ip, ipb,
  cqp40-ip, lossless-ip, lossless-ipb, cqp40-sao-ip, abr-96k;
  hevc12-* @p12: cqp-ip, cqp40-sao-ip, lossless-ip).
- Unit tests: deep_pictures_round_trip_through_the_decoder (10 and 12 bits,
  four chroma formats, I/P/B, QP 26 / QP 40 + SAO / lossless exact, SELF
  through the production decoder, with a guard that the source and the
  reconstruction really leave 8 bits); a_sample_above_the_declared_depth_
  refuses; deeper_than_fourteen_bits_refuses; deep_headers_round_trip_
  through_the_parsers (SPS depth + profile, slice QP at -QpBdOffsetY and
  51 accepted, one beyond rejected, PPS init_qp at -QpBdOffsetY).

RUN:
  cargo test --quiet: "test result: ok. 228 passed; 0 failed" (+3 doc/integration)
  verify_encode.sh (after): "encode: 315 passed, 0 failed" / "ALL GREEN"
    (280 existing cells unchanged + 35 deep cells: 4 p10 clips x 8 rows,
     1 p12 clip x 3 rows; every deep cell passed SELF, CROSS against
     libavcodec 8.1.1, and BOX; lossless rows exact)
  8-bit byte identity: hevc10_md5.sh before/after over 109 HEVC cells,
    diff empty ("BYTE-IDENTICAL: all 8-bit cells")
  MUTATIONS (see the report): A drop QpBdOffsetY in code_luma_tb;
    B writer cMax back to 7; C SPS chroma depth 0.
  cargo test --release --quiet: "test result: ok. 228 passed; 0 failed"
  cargo doc --no-deps --quiet: 0 warnings
  verify.sh --baseline baseline.txt: "baseline: every stream decodes to the
    same bytes as baseline.txt" / "ALL GREEN"
  Mutations (gate over detail10 + motion10 + the 8-bit detail control):
    A  "encode: 35 passed, 12 failed" — every lossy 10-bit cell SELF-FAIL,
       every lossless and every 8-bit cell green
    B  "encode: 46 passed, 1 failed" — motion10/hevc10-cqp40-sao-ip SELF-FAIL
       (the one deep cell whose SAO decision chose an offset above 7)
    C  "encode: 31 passed, 16 failed" — all 16 deep cells SELF-FAIL, 8-bit green

Not done: H.264 High 10 — every H.264 decision module is concretely u8
(0 uses of Sample across h264{,_intra,_me,_pic,_cavlc_mb,_cabac_mb,
_deblock}.rs, ~190 u8 sites), most of them the B-shapes track's files.
Negative quantisers (-QpBdOffsetY..-1) are spellable by the syntax layer
but not offered by Config::ConstantQp(u8), which H.264 and rivet share.
rustfmt not run: none of the touched files was rustfmt-clean at HEAD.

+++ src/encode/h265_deblock.rs
@@ -222,7 +222,7 @@ pub fn deblock_inter_picture<S: Sample>(ctx: &IntraCtx<'_, S>, pic: &mut InterPi
 /// stream will hold.
 fn run_filter<S: Sample>(ctx: &IntraCtx<'_, S>, recon: &mut Frame<S>, info: &PicInfo) {
     let h4 = recon.height / 4;
-    let pps = Pps::parse(&crate::nal::unescape_rbsp(&crate::encode::h265_syntax::write_pps(ctx.qp.clamp(0, 51) as u8, ctx.bypass, true)))
+    let pps = Pps::parse(&crate::nal::unescape_rbsp(&crate::encode::h265_syntax::write_pps(ctx.qp.clamp(0, 51), ctx.bypass, true)))
         .expect("the encoder's own PPS parses");
     let mut scratch = DeblockScratch::default();
     deblock_rows(ctx.dsp, &mut scratch, recon, info, &pps, ctx.bit_depth, ctx.bit_depth, 0, h4);
